+++ skpad-aos-sdk/SKPAdBenefitApp/src/main/java/com/skplanet/app/skpadbenefitsample/adapter/FeedAdapter.java
@@ -6,6 +6,7 @@ import com.skplanet.app.skpadbenefitsample.Constants;
 import com.skplanet.app.skpadbenefitsample.CustomFeedAdsAdapter;
 import com.skplanet.app.skpadbenefitsample.CustomFeedFeedbackHandler;
 import com.skplanet.app.skpadbenefitsample.CustomFeedHeaderViewAdapter;
+import com.skplanet.app.skpadbenefitsample.CustomFeedToolbarHolder;
 import com.skplanet.skpad.benefit.presentation.feed.FeedConfig;
 import com.skplanet.skpad.benefit.presentation.feed.FeedHandler;
 import com.skplanet.skpad.benefit.presentation.feed.header.DefaultFeedHeaderViewAdapter;
@@ -28,6 +29,7 @@ public class FeedAdapter {
         FeedConfig.Builder builder = new FeedConfig.Builder(context, Constants.FEED_UNIT_ID)
                 .feedHeaderViewAdapterClass(CustomFeedHeaderViewAdapter.class)
                 .adsAdapterClass(CustomFeedAdsAdapter.class)
+                .feedToolbarHolderClass(CustomFeedToolbarHolder.class)
                 .showInquiryButton(Constants.OLDER_14YEAR)  // // 만 14세 이상인 경우에만 VOC(문의하기) 기능을 노출해야합니다.
                 .feedFeedbackHandlerClass(CustomFeedFeedbackHandler.class);